backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "docker>=7.1.0",
    "fastapi>=0.123.0",
    "httpx>=0.28.1",
    "fakeredis>=2.23.2",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
    "sharedb>=1.0.0",
    "python-socketio>=5.10.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "greenlet>=3.2.4",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "ruff>=0.14.7",
    "types-redis>=4.6.0.20241004",
]
